Create adjustable buck model with adjustable Vset
Simulated buck @10kHz Fs with digital PID controller sampling at 2.4Mhz. The simulation indicates that
its hard to get Vo of 1.2V and below because the digital PID approach seems to oscillate the Vout a lot.

The failure to get even 1.2V Vout to regulate with digital PID prompted me to look at  different way to
have adjustable Vset and Iset. One way I thought of was to add offset into the Vfb and Ifb. The model I'm uploading explored adding offset to Vfb to achieve a variable Vset. Overall, this approach seems to
do better than the digital PID approach as it can regulate down to 1V without anyproblem. However, it can't seem to regulate down below 1V and I havent figured out why this is the case.