Upload simulation files used for compensation & some resources
Also upload a LTspice simulation used to evaluate the switching time
of the MOSFET. This was investigated because I wanted to know if the
MCU would be fast enough to sample Iout and Vout even when the duty
cycle is very very low. Conclusion is... that the MCU is not fast enough so
I am forced to do the compensation in analogue instead of digital.

Also upload some new resources. These are just some STM32 pinout
and LCD related documents. Also upload document related to good
grounding prentices for mixed signal PCBs.